refactor: condition
