+++ packages/apps/project/src/hooks/use-reporting-configuration.ts
@@ -1,0 +1,50 @@
+/*
+ * Axelor Business Solutions
+ *
+ * Copyright (C) 2024 Axelor (<http://axelor.com>).
+ *
+ * This program is free software: you can redistribute it and/or  modify
+ * it under the terms of the GNU Affero General Public License, version 3,
+ * as published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import {useMemo} from 'react';
+import {useSelector} from '@axelor/aos-mobile-core';
+import {ReportingType} from '../types';
+
+export const useReportingConfiguration = project => {
+  const {mobileSettings} = useSelector(state => state.appConfig);
+
+  const allowedReportingTypes = useMemo(() => {
+    return mobileSettings?.reportingTypesToDisplay || [];
+  }, [mobileSettings?.reportingTypesToDisplay]);
+
+  const reportingConfig = useMemo(() => {
+    const isNoneOnly =
+      allowedReportingTypes.length === 1 &&
+      allowedReportingTypes.includes(ReportingType.none);
+    const showActivities = allowedReportingTypes.includes(
+      ReportingType.activities,
+    );
+    const showReporting =
+      allowedReportingTypes.includes(ReportingType.indicator) &&
+      project?.isBusinessProject;
+
+    return {
+      isNoneOnly,
+      showActivities,
+      showReporting,
+      showReportingOrActivities: showActivities || showReporting,
+    };
+  }, [allowedReportingTypes, project]);
+
+  return reportingConfig;
+};
