+++ packages/apps/project/src/index.ts
@@ -80,4 +80,4 @@ export * from './api';
 export * from './components';
 export * from './features/asyncFunctions-index';
 export * from './screens';
-export * from './utils';
+export * from './types';
