+++ packages/apps/project/src/types/reporting-type.ts
@@ -16,4 +16,8 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-export * from './reportingUtils';
+export const ReportingType = {
+  indicator: 'indicators',
+  activities: 'activities',
+  none: 'none',
+};
